理解のためのコメント追記

+++ post/delete.php
@@ -4,9 +4,14 @@ session_start();
 
 require_once('../config/db.php');
 
-$id = $_GET['id'];
+if (!isset($_SESSION['user'])) { //ログインしていないユーザがURLから削除できないように
+    header('Location: ../auth/login.php');
+    exit;
+}
 
-$sql = "DELETE FROM posts WHERE id=? AND user_id=?";
+$id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT); // URLのクエリパラメータ（?id=○○）から投稿IDを取得
+
+$sql = "DELETE FROM posts WHERE id=? AND user_id=?"; // 投稿IDが一致し、かつログインユーザーのIDと一致する場合のみ削除
 
 $stmt = $pdo->prepare($sql);
